--- src/main/java/com/sparta/cloneproject/controller/dto/post/RequestDto.java
@@ -1,4 +1,0 @@
-package com.sparta.cloneproject.controller.dto.post;
-
-public class RequestDto {
-}
