+++ src/main/java/com/sparta/cloneproject/dto/post/RequestDto.java
@@ -1,0 +1,4 @@
+package com.sparta.cloneproject.dto.post;
+
+public class RequestDto {
+}
